New dataset chart component
