+++ asreview/webapp/src/ProjectComponents/SetupComponents/DataComponents/DatasetInfo.js
@@ -18,6 +18,7 @@ import { styled, useTheme } from "@mui/material/styles";
 import Chart from "react-apexcharts";
 
 import { projectModes } from "globals.js";
+import DatasetChart from "ProjectComponents/AnalyticsComponents/DatasetChart";
 
 
 const classes = {};
@@ -28,12 +29,12 @@ const Root = styled("div")(({ theme }) => ({}));
 const DatasetInfo = ({ project_id, dataset_path, setDataset }, props) => {
 
   console.log("render state")
-  const {
-    data,
-    error: fetchDataError,
-    isError: isFetchDataError,
-    isFetching: isFetchingData,
-  } = useQuery(
+  const progressQuery = useQuery(
+    // data,
+    // error: fetchDataError,
+    // isError: isFetchDataError,
+    // isFetching: isFetchingData,
+  // } = useQuery(
     ["fetchData", { project_id: project_id }],
     ProjectAPI.fetchData,
     {
@@ -51,506 +52,7 @@ const DatasetInfo = ({ project_id, dataset_path, setDataset }, props) => {
     },
   );
 
-  const theme = useTheme();
-  const n_papers = data?.n_rows;
-  const n_included = data?.n_relevant;
-  const n_excluded = data?.n_irrelevant;
-  const n_duplicates = data?.n_duplicates;
-  const n_english = data?.n_english;
-
-  const formattedTotal = React.useCallback(() => {
-    if (props.mode !== projectModes.SIMULATION || !props.isSimulating) {
-      return n_papers ? n_papers.toLocaleString("en-US") : 0;
-    } else {
-      return (
-        Math.round(((n_included + n_excluded) / n_papers) * 10000) / 100 + "%"
-      );
-    }
-  }, [props.isSimulating, props.mode, n_included, n_excluded, n_papers]);
-
-  const seriesArray = React.useCallback(() => {
-    if (n_papers) {
-      return [
-        Math.round(((data?.n_rows-data?.n_duplicates) / n_papers) * 10000) / 100,
-        Math.round((n_duplicates / n_papers) * 10000) / 100,
-      ];
-    } else {
-      return [];
-    }
-  }, [data?.n_duplicates, n_papers]);
-
-  const seriesArray2 = React.useCallback(() => {
-    if (n_papers) {
-      return [
-        Math.round(((data?.n_rows-data?.n_missing_title) / n_papers) * 10000) / 100,
-        Math.round((data?.n_missing_title / n_papers) * 10000) / 100,
-      ];
-    } else {
-      return [];
-    }
-  }, [data?.n_missing_title, n_papers]);
-
-  const seriesArray3 = React.useCallback(() => {
-    if (n_papers) {
-      return [
-        Math.round(((data?.n_rows-data?.n_missing_abstract) / n_papers) * 10000) / 100,
-        Math.round((data?.n_missing_abstract / n_papers) * 10000) / 100,
-      ];
-    } else {
-      return [];
-    }
-  }, [data?.n_missing_abstract, n_papers]);
-
-  const seriesArray4 = React.useCallback(() => {
-    if (n_papers) {
-      return [
-        Math.round(((n_english) / n_papers) * 10000) / 100,
-        Math.round(( (n_papers - n_english) / n_papers) * 10000) / 100,
-      ];
-    } else {
-      return [];
-    }
-  }, [n_english, n_papers]);
-
-  const optionsChart = React.useCallback(() => {
-  return {
-    chart: {
-      animations: {
-        enabled: false,
-      },
-      background: "transparent",
-      id: "ASReviewLABprogressChart",
-      type: "radialBar",
-    },
-    plotOptions: {
-      radialBar: {
-        hollow: {
-          margin: 15,
-          size: "60%",
-        },
-        dataLabels: {
-          name: {
-            fontSize: "22px",
-          },
-          value: {
-            fontSize: !props.mobileScreen
-              ? theme.typography.h5.fontSize
-              : theme.typography.h6.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.h5.fontFamily
-              : theme.typography.h6.fontFamily,
-            fontWeight: theme.typography.fontWeightBold,
-          },
-          total: {
-            show: true,
-            label:
-              props.mode !== projectModes.SIMULATION || !props.isSimulating
-                ? "Total records"
-                : "Simulation progress",
-            fontSize: !props.mobileScreen
-              ? theme.typography.subtitle1.fontSize
-              : theme.typography.subtitle2.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.subtitle1.fontFamily
-              : theme.typography.subtitle2.fontFamily,
-            color: theme.palette.text.secondary,
-            formatter: formattedTotal,
-          },
-        },
-      },
-    },
-    colors: [
-      theme.palette.mode === "light"
-        ? theme.palette.secondary.light
-        : theme.palette.secondary.main,
-      theme.palette.mode === "light"
-        ? theme.palette.primary.light
-        : theme.palette.primary.main,
-    ],
-    dataLabels: {
-      enabled: false,
-    },
-    labels: ["Unique", "Duplicate"],
-    legend: {
-      show: true,
-      position: "bottom",
-      fontSize: !props.mobileScreen ? "14px" : "12px",
-      fontFamily: theme.typography.subtitle2.fontFamily,
-      fontWeight: theme.typography.subtitle2.fontWeight,
-      labels: {
-        colors: theme.palette.text.secondary,
-      },
-      markers: {
-        width: 8,
-        height: 8,
-        offsetX: -4,
-      },
-      itemMargin: {
-        horizontal: 16,
-      },
-    },
-    fill: {
-      type: "gradient",
-      gradient: {
-        shade: "light",
-        type: "horizontal",
-        shadeIntensity: 0,
-        inverseColors: true,
-        opacityFrom: 0.7,
-        opacityTo: 0.9,
-        stops: [0, 100],
-      },
-    },
-    markers: {
-      size: 0,
-    },
-    noData: {
-      text: "No data available",
-    },
-    stroke: {
-      lineCap: "round",
-    },
-    theme: {
-      mode: theme.palette.mode,
-    },
-  };
-});
-
-const optionsChart2 = React.useCallback(() => {
-  return {
-    chart: {
-      animations: {
-        enabled: false,
-      },
-      background: "transparent",
-      id: "ASReviewLABprogressChart",
-      type: "radialBar",
-    },
-    plotOptions: {
-      radialBar: {
-        hollow: {
-          margin: 15,
-          size: "60%",
-        },
-        dataLabels: {
-          name: {
-            fontSize: "22px",
-          },
-          value: {
-            fontSize: !props.mobileScreen
-              ? theme.typography.h5.fontSize
-              : theme.typography.h6.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.h5.fontFamily
-              : theme.typography.h6.fontFamily,
-            fontWeight: theme.typography.fontWeightBold,
-          },
-          total: {
-            show: true,
-            label:
-              props.mode !== projectModes.SIMULATION || !props.isSimulating
-                ? "Total records"
-                : "Simulation progress",
-            fontSize: !props.mobileScreen
-              ? theme.typography.subtitle1.fontSize
-              : theme.typography.subtitle2.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.subtitle1.fontFamily
-              : theme.typography.subtitle2.fontFamily,
-            color: theme.palette.text.secondary,
-            formatter: formattedTotal,
-          },
-        },
-      },
-    },
-    colors: [
-      theme.palette.mode === "light"
-        ? theme.palette.secondary.light
-        : theme.palette.secondary.main,
-      theme.palette.mode === "light"
-        ? theme.palette.primary.light
-        : theme.palette.primary.main,
-    ],
-    dataLabels: {
-      enabled: false,
-    },
-    labels: ["Has title", "No title"],
-    legend: {
-      show: true,
-      position: "bottom",
-      fontSize: !props.mobileScreen ? "14px" : "12px",
-      fontFamily: theme.typography.subtitle2.fontFamily,
-      fontWeight: theme.typography.subtitle2.fontWeight,
-      labels: {
-        colors: theme.palette.text.secondary,
-      },
-      markers: {
-        width: 8,
-        height: 8,
-        offsetX: -4,
-      },
-      itemMargin: {
-        horizontal: 16,
-      },
-    },
-    fill: {
-      type: "gradient",
-      gradient: {
-        shade: "light",
-        type: "horizontal",
-        shadeIntensity: 0,
-        inverseColors: true,
-        opacityFrom: 0.7,
-        opacityTo: 0.9,
-        stops: [0, 100],
-      },
-    },
-    markers: {
-      size: 0,
-    },
-    noData: {
-      text: "No data available",
-    },
-    stroke: {
-      lineCap: "round",
-    },
-    theme: {
-      mode: theme.palette.mode,
-    },
-  };
-});
-
-const optionsChart3 = React.useCallback(() => {
-  return {
-    chart: {
-      animations: {
-        enabled: false,
-      },
-      background: "transparent",
-      id: "ASReviewLABprogressChart",
-      type: "radialBar",
-    },
-    plotOptions: {
-      radialBar: {
-        hollow: {
-          margin: 15,
-          size: "60%",
-        },
-        dataLabels: {
-          name: {
-            fontSize: "22px",
-          },
-          value: {
-            fontSize: !props.mobileScreen
-              ? theme.typography.h5.fontSize
-              : theme.typography.h6.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.h5.fontFamily
-              : theme.typography.h6.fontFamily,
-            fontWeight: theme.typography.fontWeightBold,
-          },
-          total: {
-            show: true,
-            label:
-              props.mode !== projectModes.SIMULATION || !props.isSimulating
-                ? "Total records"
-                : "Simulation progress",
-            fontSize: !props.mobileScreen
-              ? theme.typography.subtitle1.fontSize
-              : theme.typography.subtitle2.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.subtitle1.fontFamily
-              : theme.typography.subtitle2.fontFamily,
-            color: theme.palette.text.secondary,
-            formatter: formattedTotal,
-          },
-        },
-      },
-    },
-    colors: [
-      theme.palette.mode === "light"
-        ? theme.palette.secondary.light
-        : theme.palette.secondary.main,
-      theme.palette.mode === "light"
-        ? theme.palette.primary.light
-        : theme.palette.primary.main,
-    ],
-    dataLabels: {
-      enabled: false,
-    },
-    labels: ["Has abstract", "No abstract"],
-    legend: {
-      show: true,
-      position: "bottom",
-      fontSize: !props.mobileScreen ? "14px" : "12px",
-      fontFamily: theme.typography.subtitle2.fontFamily,
-      fontWeight: theme.typography.subtitle2.fontWeight,
-      labels: {
-        colors: theme.palette.text.secondary,
-      },
-      markers: {
-        width: 8,
-        height: 8,
-        offsetX: -4,
-      },
-      itemMargin: {
-        horizontal: 16,
-      },
-    },
-    fill: {
-      type: "gradient",
-      gradient: {
-        shade: "light",
-        type: "horizontal",
-        shadeIntensity: 0,
-        inverseColors: true,
-        opacityFrom: 0.7,
-        opacityTo: 0.9,
-        stops: [0, 100],
-      },
-    },
-    markers: {
-      size: 0,
-    },
-    noData: {
-      text: "No data available",
-    },
-    stroke: {
-      lineCap: "round",
-    },
-    theme: {
-      mode: theme.palette.mode,
-    },
-  };
-});
-
-const optionsChart4 = React.useCallback(() => {
-  return {
-    chart: {
-      animations: {
-        enabled: false,
-      },
-      background: "transparent",
-      id: "ASReviewLABprogressChart",
-      type: "radialBar",
-    },
-    plotOptions: {
-      radialBar: {
-        hollow: {
-          margin: 15,
-          size: "60%",
-        },
-        dataLabels: {
-          name: {
-            fontSize: "22px",
-          },
-          value: {
-            fontSize: !props.mobileScreen
-              ? theme.typography.h5.fontSize
-              : theme.typography.h6.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.h5.fontFamily
-              : theme.typography.h6.fontFamily,
-            fontWeight: theme.typography.fontWeightBold,
-          },
-          total: {
-            show: true,
-            label:
-              props.mode !== projectModes.SIMULATION || !props.isSimulating
-                ? "Total records"
-                : "Simulation progress",
-            fontSize: !props.mobileScreen
-              ? theme.typography.subtitle1.fontSize
-              : theme.typography.subtitle2.fontSize,
-            fontFamily: !props.mobileScreen
-              ? theme.typography.subtitle1.fontFamily
-              : theme.typography.subtitle2.fontFamily,
-            color: theme.palette.text.secondary,
-            formatter: formattedTotal,
-          },
-        },
-      },
-    },
-    colors: [
-      theme.palette.mode === "light"
-        ? theme.palette.secondary.light
-        : theme.palette.secondary.main,
-      theme.palette.mode === "light"
-        ? theme.palette.primary.light
-        : theme.palette.primary.main,
-    ],
-    dataLabels: {
-      enabled: false,
-    },
-    labels: ["English", "Non English"],
-    legend: {
-      show: true,
-      position: "bottom",
-      fontSize: !props.mobileScreen ? "14px" : "12px",
-      fontFamily: theme.typography.subtitle2.fontFamily,
-      fontWeight: theme.typography.subtitle2.fontWeight,
-      labels: {
-        colors: theme.palette.text.secondary,
-      },
-      markers: {
-        width: 8,
-        height: 8,
-        offsetX: -4,
-      },
-      itemMargin: {
-        horizontal: 16,
-      },
-    },
-    fill: {
-      type: "gradient",
-      gradient: {
-        shade: "light",
-        type: "horizontal",
-        shadeIntensity: 0,
-        inverseColors: true,
-        opacityFrom: 0.7,
-        opacityTo: 0.9,
-        stops: [0, 100],
-      },
-    },
-    markers: {
-      size: 0,
-    },
-    noData: {
-      text: "No data available",
-    },
-    stroke: {
-      lineCap: "round",
-    },
-    theme: {
-      mode: theme.palette.mode,
-    },
-  };
-});
-
-  const [series, setSeries] = React.useState(seriesArray());
-  const [options, setOptions] = React.useState({});
-  const [options2, setOptions2] = React.useState({});
-  const [series2, setSeries2] = React.useState(seriesArray2());
-  const [options3, setOptions3] = React.useState({});
-  const [series3, setSeries3] = React.useState(seriesArray3());
-  const [options4, setOptions4] = React.useState({});
-  const [series4, setSeries4] = React.useState(seriesArray4());
-
-  React.useEffect(() => {
-    setSeries(seriesArray());
-    setOptions(optionsChart());
-
-    setSeries2(seriesArray2());
-    setOptions2(optionsChart2());
-
-    setSeries3(seriesArray3());
-    setOptions3(optionsChart3());
-
-    setSeries4(seriesArray4());
-    setOptions4(optionsChart4());
-    
-  }, [seriesArray]);//[seriesArray, optionsChart, seriesArray2, optionsChart2, seriesArray3, optionsChart3 , seriesArray4, optionsChart4]);
+  const n_english = 30;
 
   return (
     <Root>
@@ -580,7 +82,7 @@ const optionsChart4 = React.useCallback(() => {
               <Typography variant="body2" sx={{ color: "text.secondary" }}>
                 Records
               </Typography>
-              <Typography variant="body2">{data?.n_rows}</Typography>
+              <Typography variant="body2">{progressQuery.data?.n_rows}</Typography>
             </Stack>
 
             <Stack>
@@ -588,7 +90,7 @@ const optionsChart4 = React.useCallback(() => {
                 Duplicates
               </Typography>
               <Typography variant="body2">
-                About {data?.n_duplicates} ({(Math.round((data?.n_duplicates/data?.n_rows*100)*100)/100).toFixed(2)}%)
+                About {progressQuery.data?.n_duplicates} ({(Math.round((progressQuery.data?.n_duplicates/progressQuery.data?.n_rows*100)*100)/100).toFixed(2)}%)
               </Typography>
               </Stack>
 
@@ -597,7 +99,7 @@ const optionsChart4 = React.useCallback(() => {
                 Missing titles
               </Typography>
               <Typography variant="body2">
-                {data?.n_missing_title}
+                {progressQuery.data?.n_missing_title}
               </Typography>
               </Stack>
 
@@ -606,54 +108,66 @@ const optionsChart4 = React.useCallback(() => {
                 Missing abstracts
               </Typography>
               <Typography variant="body2">
-                {data?.n_missing_abstract}
+                {progressQuery.data?.n_missing_abstract}
               </Typography>
               </Stack>
 
-            <Box style={{display: 'flex', justifyContent: 'space-between'}}>
-              <Chart
-          options={options}
-          series={series}
-          type="radialBar"
-          height={350}
-          style={{flex: 1}}
+        {/* <Box style={{display: 'flex', justifyContent: 'space-between'}}> */}
+        <Box style={{display: 'flex', justifyContent: 'space-between'}}>
+        <DatasetChart 
+        isSimulating={props.isSimulating}
+        mobileScreen={props.mobileScreen}
+        mode={props.mode}
+        progressQuery={progressQuery}
+        labels={["Unique", "Duplicate"]}
+        part={progressQuery.data?.n_rows-progressQuery.data?.n_duplicates}
+        total={progressQuery.data?.n_rows}
         />
 
-<Chart
-          options={options2}
-          series={series2}
-          type="radialBar"
-          height={350}
-          style={{flex: 1}}
+        <DatasetChart 
+        isSimulating={props.isSimulating}
+        mobileScreen={props.mobileScreen}
+        mode={props.mode}
+        progressQuery={progressQuery}
+        labels={["Has title", "No title"]}
+        part={progressQuery.data?.n_rows-progressQuery.data?.n_missing_title}
+        total={progressQuery.data?.n_rows}
         />
 
-<Chart
-          options={options3}
-          series={series3}
-          type="radialBar"
-          height={350}
-          style={{flex: 1}}
+        <DatasetChart 
+        isSimulating={props.isSimulating}
+        mobileScreen={props.mobileScreen}
+        mode={props.mode}
+        progressQuery={progressQuery}
+        labels={["Has abstract", "No abstract"]}
+        part={progressQuery.data?.n_rows-progressQuery.data?.n_missing_abstract}
+        total={progressQuery.data?.n_rows}
+        />  
+
+        <DatasetChart 
+        isSimulating={props.isSimulating}
+        mobileScreen={props.mobileScreen}
+        mode={props.mode}
+        progressQuery={progressQuery}
+        labels={["English", "Non English"]}
+        part={n_english}
+        total={progressQuery.data?.n_rows}
         />
-<Chart
-          options={options4}
-          series={series4}
-          type="radialBar"
-          height={350}
-          style={{flex: 1}}
-        />
+
         </Box>
+
 
           </Stack>
 
-          {isFetchingData && (
+          {progressQuery.isFetchingData && (
             <Box className="main-page-body-wrapper">
               <CircularProgress />
             </Box>
           )}
           <CardErrorHandler
             queryKey={"fetchData"}
-            error={fetchDataError}
-            isError={isFetchDataError}
+            error={progressQuery.fetchDataError}
+            isError={progressQuery.isFetchDataError}
           />
 
           <Button
